+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+
+- Add @@clear-and-update-index-server-index. [ksuess]
+- Fix error handling in REST API service @kitsearch [ksuess]
+
+
 2.0.1 (2024-02-05)
 ------------------